+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Bug Fixes
+
+- fix(apple): Fix duplicate `init()` being added to SwiftUI apps that already have an initializer
+
 ## 6.11.0
 
 ### New Features